Used async (non-thread-based) calls to do direct updates.
This strategy resulted in much faster --update-direct numbers. Now
we're seeing about 75 updates per second for 100k docs where as the old
multi-threaded calls yielded around half that.

The async style calls are implemented using Scala's Actor model, which
is truly a pleasure to code with! Being able to declare tasks
asyncrhonously is so much easier in Scala.